The JS UI calls a ‘saveEdit()’ function which grabs all the information from the StoryView and sends an HTTP PUT.
For some reason, instead of sending a parameter called ‘documents’ with just an array of JSON objects, it’s actually sending an Array of hash objects, all with a ‘file’ key pointing to the actual document.

In the stories_controllers, it sets ‘story.documents = allowed_params[:documents]’), and because the Safe Params rules out the ‘file’ key, it sets the documents association with an array of empty hashes, and this cleans out the documents association.

The ‘quick fix’ is to fix the params.

It’s not optimal because every drag&drop actually deletes the files from the ‘attachinary_files’ table and recreates them, but at least it’s better than losing the associations

#MUST_MAKE_BETTER