Adding a new capybara test environment to force assets precompilation, to make sure all feature specs use the exact same static assets. This should also make feature specs faster (10 to 20 seconds faster in total).
One failing spec was the notes feature spec because the action of adding a new note had to wait for the setInterval to project.fetch the last changeset in order to refresh the story view. Now the add note success hook will forcefully fetch the last changeset, making it appear “faster” and also avoiding timeout in the feature spec.

The iteration service spec relied on fixing the random seed to generate the same set of stories with estimates. But sometimes it failed in the ordering. Instead, we are fixing a set of random numbers in an array to make sure every test runs exactly the same, without relying on the Random.new(666).rand

These changes “should” make it unnecessary to fix the rspec random seed.